fix issues with the java bytecode frontend
Moved the logic for assigning the values of operands to stack locals out of `OperandStack` and into `Operand`.
That means that the assignments are created when the operands are used as a local/immediate with the `toLocal` and `toImmediate` methods instead of when they are popped of the stack with `popLocal` and `popImmediate`.
This fixes some issues where the wrong `pop` method was used which caused unnecessary stack local variables to get created.
This also results in improved type-safety and gets rid of a bunch of casts, because `toLocal` and `toImmediate` actually return the `Local` and `Immediate` types instead of a generic `Operand`.

Significantly simplified merging logic in `StackFrame`.
Factored a bunch of duplicated code into the new `changeStackLocal` in `Operand`.
The old version actually incorrectly merged operands at times (See `SwitchExprWithYieldTest`). I will add a test for this in a later commit.

Fixed updating usages of the `value`/`stackLocal` in `Operand`. This is now part of the `changeStackLocal` method.

Simplified duplication instructions (Thanks to `Operand.DWORD_DUMMY`).

Removed some commented out code.