fix position info for statements
When converting the instructions for a branch, there might not be a line number node after the branch target.
This would result in incorrect line numbers being used until the next line number node.

The fix is to simply store the line number for branches and restore the current line number when handling that branch.