handle typing of locals with duplicate names
Since locals with the same name can't be differentiated in Jimple, they need to be treated the same, even when they are actually different locals.

This is currently only possible because the bytecode to Jimple conversion tries to use the provided debug names of locals, but different locals (meaning local ids) might have the same name.