introduce a `TopType`
The `TypeAssigner`/`TypeResolver` previously wouldn't function correctly when a variable got assigned both primitive and non-primitive values. The top type fixes this.

The top type is a superclass of all other types. This is similar to `java.lang.Object` but
also includes primitive types.

This type can't exist in Java source code, but it can implicitly exist in bytecode. This happens
when the compiler re-uses local variables with the same id, but different types.

If you see this type when you didn't expect it, you probably need to **turn on the `LocalSplitter`**. The `LocalSplitter` will remove all situations where a `TopType` could be created by
the `TypeAssigner` (at least when the bytecode has been generated from Java source code).