fix evaluation of array reference expressions
Consider the statement `l1 = l2[l3]`, which contains the expression `l2[l3]`.

When `l2` has a non-array type, it can't be known what the type of the array ref expression is. Because the result of the array access could be an object or a primitive, the top type has to be chosen here.