sootup.core: fix some compiler deprecation warnings
When compiling sootup.core with -Xlint:unchecked, the following warnings
are reported:

[WARNING] /home/marcus/sootUp/sootup.core/src/main/java/sootup/core/util/DotExporter.java:[205,12] org.apache.commons.lang3.StringEscapeUtils in org.apache.commons.lang3 has been deprecated
[WARNING] /home/marcus/sootUp/sootup.core/src/main/java/sootup/core/util/Utils.java:[236,25] org.apache.commons.lang3.StringEscapeUtils in org.apache.commons.lang3 has been deprecated

According to the javadocs of org.apache.commons.lang3.StringEscapeUtils, the
StringEscapeUtils from the org.apache.commons.text package should be used
instead (according to the docs, the text.StringEscapeUtils.escapeXml10 method
behaves exactly as the lang3.StringEscapeUtils.escapeXml10 method).
Unfortunately, this introduces a new dependency to the "commons-text" artifact.